+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/query/LogQueryService.java
@@ -22,6 +22,8 @@ import java.io.IOException;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.apache.skywalking.oap.server.core.query.input.Duration;
+import org.apache.skywalking.oap.server.core.query.type.debugging.DebuggingSpan;
+import org.apache.skywalking.oap.server.core.query.type.debugging.DebuggingTraceContext;
 import org.apache.skywalking.oap.server.library.util.StringUtil;
 import org.apache.skywalking.oap.server.core.analysis.IDManager;
 import org.apache.skywalking.oap.server.core.analysis.manual.searchtag.Tag;
@@ -35,6 +37,7 @@ import org.apache.skywalking.oap.server.library.module.ModuleManager;
 import org.apache.skywalking.oap.server.library.module.Service;
 
 import static java.util.Objects.nonNull;
+import static org.apache.skywalking.oap.server.core.query.type.debugging.DebuggingTraceContext.TRACE_CONTEXT;
 
 public class LogQueryService implements Service {
 
@@ -66,6 +69,45 @@ public class LogQueryService implements Service {
                           final List<Tag> tags,
                           List<String> keywordsOfContent,
                           List<String> excludingKeywordsOfContent) throws IOException {
+        DebuggingTraceContext traceContext = TRACE_CONTEXT.get();
+        DebuggingSpan span = null;
+        try {
+            if (traceContext != null) {
+                StringBuilder msg = new StringBuilder();
+                span = traceContext.createSpan("Query Service: queryLogs");
+                msg.append("ServiceId: ").append(serviceId).append(", ");
+                msg.append("ServiceInstanceId: ").append(serviceInstanceId).append(", ");
+                msg.append("EndpointId: ").append(endpointId).append(", ");
+                msg.append("RelatedTrace: ").append(relatedTrace).append(", ");
+                msg.append("Pagination: ").append(paging).append(", ");
+                msg.append("QueryOrder: ").append(queryOrder).append(", ");
+                msg.append("Duration: ").append(duration).append(", ");
+                msg.append("Tags: ").append(tags).append(", ");
+                msg.append("KeywordsOfContent: ").append(keywordsOfContent).append(", ");
+                msg.append("ExcludingKeywordsOfContent: ").append(excludingKeywordsOfContent);
+                span.setMsg(msg.toString());
+            }
+            return invokeQueryLogs(
+                serviceId, serviceInstanceId, endpointId, relatedTrace, paging, queryOrder, duration, tags,
+                keywordsOfContent, excludingKeywordsOfContent
+            );
+        } finally {
+            if (traceContext != null) {
+                traceContext.stopSpan(span);
+            }
+        }
+    }
+
+    private Logs invokeQueryLogs(String serviceId,
+                          String serviceInstanceId,
+                          String endpointId,
+                          TraceScopeCondition relatedTrace,
+                          Pagination paging,
+                          Order queryOrder,
+                          final Duration duration,
+                          final List<Tag> tags,
+                          List<String> keywordsOfContent,
+                          List<String> excludingKeywordsOfContent) throws IOException {
         PaginationUtils.Page page = PaginationUtils.INSTANCE.exchange(paging);
 
         if (nonNull(keywordsOfContent)) {
@@ -79,7 +121,7 @@ public class LogQueryService implements Service {
                                                                    .collect(Collectors.toList());
         }
 
-        Logs logs = getLogQueryDAO().queryLogs(serviceId,
+        Logs logs = getLogQueryDAO().queryLogsDebuggable(serviceId,
                                                serviceInstanceId,
                                                endpointId,
                                                relatedTrace,