+++ oap-server/server-query-plugin/query-graphql-plugin/src/main/java/org/apache/skywalking/oap/query/graphql/resolver/LogQuery.java
@@ -30,11 +30,14 @@ import org.apache.skywalking.oap.server.core.query.enumeration.Order;
 import org.apache.skywalking.oap.server.core.query.input.Duration;
 import org.apache.skywalking.oap.server.core.query.input.LogQueryCondition;
 import org.apache.skywalking.oap.server.core.query.type.Logs;
+import org.apache.skywalking.oap.server.core.query.type.debugging.DebuggingSpan;
+import org.apache.skywalking.oap.server.core.query.type.debugging.DebuggingTraceContext;
 import org.apache.skywalking.oap.server.library.module.ModuleManager;
 import org.apache.skywalking.oap.server.library.util.CollectionUtils;
 import org.apache.skywalking.oap.server.library.util.StringUtil;
 
 import static java.util.Objects.isNull;
+import static org.apache.skywalking.oap.server.core.query.type.debugging.DebuggingTraceContext.TRACE_CONTEXT;
 
 public class LogQuery implements GraphQLQueryResolver {
     private final ModuleManager moduleManager;
@@ -63,7 +66,24 @@ public class LogQuery implements GraphQLQueryResolver {
         return getQueryService().supportQueryLogsByKeywords();
     }
 
-    public Logs queryLogs(LogQueryCondition condition) throws IOException {
+    public Logs queryLogs(LogQueryCondition condition, boolean debug) throws IOException {
+        DebuggingTraceContext traceContext = new DebuggingTraceContext("LogQueryCondition: " + condition, debug, false);
+        DebuggingTraceContext.TRACE_CONTEXT.set(traceContext);
+        DebuggingSpan span = traceContext.createSpan("Query logs");
+        try {
+            Logs logs = invokeQueryLogs(condition);
+            if (debug) {
+                logs.setDebuggingTrace(traceContext.getExecTrace());
+            }
+            return logs;
+        } finally {
+            traceContext.stopSpan(span);
+            traceContext.stopTrace();
+            TRACE_CONTEXT.remove();
+        }
+    }
+
+    private Logs invokeQueryLogs(LogQueryCondition condition) throws IOException {
         if (isNull(condition.getQueryDuration()) && isNull(condition.getRelatedTrace())) {
             throw new UnexpectedException("The condition must contains either queryDuration or relatedTrace.");
         }